+++ src/tool/cpp/cppwinrt/strings/base_activation.h
@@ -309,6 +309,34 @@ namespace winrt::impl
 
         return { result, take_ownership_from_abi };
     }
+
+    template <> struct abi<Windows::Foundation::IActivationFactory>
+    {
+        struct WINRT_NOVTABLE type : inspectable_abi
+        {
+            virtual int32_t WINRT_CALL ActivateInstance(void** instance) noexcept = 0;
+        };
+    };
+
+    template <> struct guid_storage<Windows::Foundation::IActivationFactory>
+    {
+        static constexpr guid value{ 0x00000035,0x0000,0x0000,{ 0xc0,0x00,0x00,0x00,0x00,0x00,0x00,0x46 } };
+    };
+
+    template <typename D> struct produce<D, Windows::Foundation::IActivationFactory> : produce_base<D, Windows::Foundation::IActivationFactory>
+    {
+        int32_t WINRT_CALL ActivateInstance(void** instance) noexcept final
+        {
+            try
+            {
+                *instance = nullptr;
+                typename D::abi_guard guard(this->shim());
+                *instance = detach_abi(this->shim().ActivateInstance());
+                return error_ok;
+            }
+            catch (...) { return to_hresult(); }
+        }
+    };
 }
 
 WINRT_EXPORT namespace winrt
@@ -369,5 +397,22 @@ WINRT_EXPORT namespace winrt
         void* result;
         check_hresult(WINRT_CoCreateInstance(clsid, outer, context, guid_of<Interface>(), &result));
         return { result, take_ownership_from_abi };
+    }
+
+    namespace Windows::Foundation
+    {
+        struct IActivationFactory : IInspectable
+        {
+            IActivationFactory(std::nullptr_t = nullptr) noexcept {}
+            IActivationFactory(void* ptr, take_ownership_from_abi_t) noexcept : IInspectable(ptr, take_ownership_from_abi) {}
+
+            template <typename T>
+            T ActivateInstance() const
+            {
+                IInspectable instance;
+                check_hresult((*(impl::abi_t<IActivationFactory>**)this)->ActivateInstance(put_abi(instance)));
+                return instance.try_as<T>();
+            }
+        };
     }
 }
